requirements.txt
pandas==2.3.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.1.0